osd/HitSet: split to reduce header dependencies
`HitSet.h` is a header that is included implicitly by many other
headers via `osd_types.h`.  Splitting it allows dropping lots of heavy
dependencies like `bloom_filter.h`.
