+++ src/osd/HitSet.h
@@ -17,12 +17,10 @@
 #define CEPH_OSD_HITSET_H
 
 #include <string_view>
-#include <unordered_set>
 
 #include <boost/scoped_ptr.hpp>
 
 #include "include/encoding.h"
-#include "common/bloom_filter.hpp"
 #include "common/hobject.h"
 
 /**
@@ -175,294 +173,5 @@ WRITE_CLASS_ENCODER(HitSet::Params)
 typedef boost::shared_ptr<HitSet> HitSetRef;
 
 std::ostream& operator<<(std::ostream& out, const HitSet::Params& p);
-
-/**
- * explicitly enumerate hash hits in the set
- */
-class ExplicitHashHitSet : public HitSet::Impl {
-  uint64_t count;
-  std::unordered_set<uint32_t> hits;
-public:
-  class Params : public HitSet::Params::Impl {
-  public:
-    HitSet::impl_type_t get_type() const override {
-      return HitSet::TYPE_EXPLICIT_HASH;
-    }
-    HitSet::Impl *get_new_impl() const override {
-      return new ExplicitHashHitSet;
-    }
-    static std::list<Params> generate_test_instances() {
-      std::list<Params> o;
-      o.emplace_back();
-      return o;
-    }
-  };
-
-  ExplicitHashHitSet() : count(0) {}
-  explicit ExplicitHashHitSet(const ExplicitHashHitSet::Params *p) : count(0) {}
-  ExplicitHashHitSet(const ExplicitHashHitSet &o) : count(o.count),
-      hits(o.hits) {}
-
-  HitSet::Impl *clone() const override {
-    return new ExplicitHashHitSet(*this);
-  }
-
-  HitSet::impl_type_t get_type() const override {
-    return HitSet::TYPE_EXPLICIT_HASH;
-  }
-  bool is_full() const override {
-    return false;
-  }
-  void insert(const hobject_t& o) override {
-    hits.insert(o.get_hash());
-    ++count;
-  }
-  bool contains(const hobject_t& o) const override {
-    return hits.count(o.get_hash());
-  }
-  unsigned insert_count() const override {
-    return count;
-  }
-  unsigned approx_unique_insert_count() const override {
-    return hits.size();
-  }
-  void encode(ceph::buffer::list &bl) const override {
-    ENCODE_START(1, 1, bl);
-    encode(count, bl);
-    encode(hits, bl);
-    ENCODE_FINISH(bl);
-  }
-  void decode(ceph::buffer::list::const_iterator &bl) override {
-    DECODE_START(1, bl);
-    decode(count, bl);
-    decode(hits, bl);
-    DECODE_FINISH(bl);
-  }
-  void dump(ceph::Formatter *f) const override;
-  static std::list<ExplicitHashHitSet> generate_test_instances() {
-    std::list<ExplicitHashHitSet> o;
-    o.emplace_back();
-    o.emplace_back();
-    o.back().insert(hobject_t());
-    o.back().insert(hobject_t("asdf", "", CEPH_NOSNAP, 123, 1, ""));
-    o.back().insert(hobject_t("qwer", "", CEPH_NOSNAP, 456, 1, ""));
-    return o;
-  }
-};
-WRITE_CLASS_ENCODER(ExplicitHashHitSet)
-
-/**
- * explicitly enumerate objects in the set
- */
-class ExplicitObjectHitSet : public HitSet::Impl {
-  uint64_t count;
-  std::unordered_set<hobject_t> hits;
-public:
-  class Params : public HitSet::Params::Impl {
-  public:
-    HitSet::impl_type_t get_type() const override {
-      return HitSet::TYPE_EXPLICIT_OBJECT;
-    }
-    HitSet::Impl *get_new_impl() const override {
-      return new ExplicitObjectHitSet;
-    }
-    static std::list<Params> generate_test_instances() {
-      std::list<Params> o;
-      o.emplace_back();
-      return o;
-    }
-  };
-
-  ExplicitObjectHitSet() : count(0) {}
-  explicit ExplicitObjectHitSet(const ExplicitObjectHitSet::Params *p) : count(0) {}
-  ExplicitObjectHitSet(const ExplicitObjectHitSet &o) : count(o.count),
-      hits(o.hits) {}
-
-  HitSet::Impl *clone() const override {
-    return new ExplicitObjectHitSet(*this);
-  }
-
-  HitSet::impl_type_t get_type() const override {
-    return HitSet::TYPE_EXPLICIT_OBJECT;
-  }
-  bool is_full() const override {
-    return false;
-  }
-  void insert(const hobject_t& o) override {
-    hits.insert(o);
-    ++count;
-  }
-  bool contains(const hobject_t& o) const override {
-    return hits.count(o);
-  }
-  unsigned insert_count() const override {
-    return count;
-  }
-  unsigned approx_unique_insert_count() const override {
-    return hits.size();
-  }
-  void encode(ceph::buffer::list &bl) const override {
-    ENCODE_START(1, 1, bl);
-    encode(count, bl);
-    encode(hits, bl);
-    ENCODE_FINISH(bl);
-  }
-  void decode(ceph::buffer::list::const_iterator& bl) override {
-    DECODE_START(1, bl);
-    decode(count, bl);
-    decode(hits, bl);
-    DECODE_FINISH(bl);
-  }
-  void dump(ceph::Formatter *f) const override;
-  static std::list<ExplicitObjectHitSet> generate_test_instances() {
-    std::list<ExplicitObjectHitSet> o;
-    o.emplace_back();
-    o.emplace_back();
-    o.back().insert(hobject_t());
-    o.back().insert(hobject_t("asdf", "", CEPH_NOSNAP, 123, 1, ""));
-    o.back().insert(hobject_t("qwer", "", CEPH_NOSNAP, 456, 1, ""));
-    return o;
-  }
-};
-WRITE_CLASS_ENCODER(ExplicitObjectHitSet)
-
-/**
- * use a bloom_filter to track hits to the set
- */
-class BloomHitSet : public HitSet::Impl {
-  compressible_bloom_filter bloom;
-
-public:
-  HitSet::impl_type_t get_type() const override {
-    return HitSet::TYPE_BLOOM;
-  }
-
-  class Params : public HitSet::Params::Impl {
-  public:
-    HitSet::impl_type_t get_type() const override {
-      return HitSet::TYPE_BLOOM;
-    }
-    HitSet::Impl *get_new_impl() const override {
-      return new BloomHitSet;
-    }
-
-    uint32_t fpp_micro;    ///< false positive probability / 1M
-    uint64_t target_size;  ///< number of unique insertions we expect to this HitSet
-    uint64_t seed;         ///< seed to use when initializing the bloom filter
-
-    Params()
-      : fpp_micro(0), target_size(0), seed(0) {}
-    Params(double fpp, uint64_t t, uint64_t s)
-      : fpp_micro(fpp * 1000000.0), target_size(t), seed(s) {}
-    Params(const Params &o)
-      : fpp_micro(o.fpp_micro),
-	target_size(o.target_size),
-	seed(o.seed) {}
-    ~Params() override {}
-
-    double get_fpp() const {
-      return (double)fpp_micro / 1000000.0;
-    }
-    void set_fpp(double f) {
-      fpp_micro = (unsigned)(llrintl(f * 1000000.0));
-    }
-
-    void encode(ceph::buffer::list& bl) const override {
-      ENCODE_START(1, 1, bl);
-      encode(fpp_micro, bl);
-      encode(target_size, bl);
-      encode(seed, bl);
-      ENCODE_FINISH(bl);
-    }
-    void decode(ceph::buffer::list::const_iterator& bl) override {
-      DECODE_START(1, bl);
-      decode(fpp_micro, bl);
-      decode(target_size, bl);
-      decode(seed, bl);
-      DECODE_FINISH(bl);
-    }
-    void dump(ceph::Formatter *f) const override;
-    void dump_stream(std::ostream& o) const override {
-      o << "false_positive_probability: "
-	<< get_fpp() << ", target_size: " << target_size
-	<< ", seed: " << seed;
-    }
-    static std::list<Params> generate_test_instances() {
-      std::list<Params> o;
-      o.emplace_back();
-      o.emplace_back();
-      o.back().fpp_micro = 123456;
-      o.back().target_size = 300;
-      o.back().seed = 99;
-      return o;
-    }
-  };
-
-  BloomHitSet() {}
-  BloomHitSet(unsigned inserts, double fpp, int seed)
-    : bloom(inserts, fpp, seed)
-  {}
-  explicit BloomHitSet(const BloomHitSet::Params *p) : bloom(p->target_size,
-                                                    p->get_fpp(),
-                                                    p->seed)
-  {}
-
-  BloomHitSet(const BloomHitSet &o) {
-    // oh god
-    ceph::buffer::list bl;
-    o.encode(bl);
-    auto bli = std::cbegin(bl);
-    this->decode(bli);
-  }
-
-  HitSet::Impl *clone() const override {
-    return new BloomHitSet(*this);
-  }
-
-  bool is_full() const override {
-    return bloom.is_full();
-  }
-
-  void insert(const hobject_t& o) override {
-    bloom.insert(o.get_hash());
-  }
-  bool contains(const hobject_t& o) const override {
-    return bloom.contains(o.get_hash());
-  }
-  unsigned insert_count() const override {
-    return bloom.element_count();
-  }
-  unsigned approx_unique_insert_count() const override {
-    return bloom.approx_unique_element_count();
-  }
-  void seal() override {
-    // aim for a density of .5 (50% of bit set)
-    double pc = bloom.density() * 2.0;
-    if (pc < 1.0)
-      bloom.compress(pc);
-  }
-
-  void encode(ceph::buffer::list &bl) const override {
-    ENCODE_START(1, 1, bl);
-    encode(bloom, bl);
-    ENCODE_FINISH(bl);
-  }
-  void decode(ceph::buffer::list::const_iterator& bl) override {
-    DECODE_START(1, bl);
-    decode(bloom, bl);
-    DECODE_FINISH(bl);
-  }
-  void dump(ceph::Formatter *f) const override;
-  static std::list<BloomHitSet> generate_test_instances() {
-    std::list<BloomHitSet> o;
-    o.emplace_back();
-    o.push_back(BloomHitSet(10, .1, 1));
-    o.back().insert(hobject_t());
-    o.back().insert(hobject_t("asdf", "", CEPH_NOSNAP, 123, 1, ""));
-    o.back().insert(hobject_t("qwer", "", CEPH_NOSNAP, 456, 1, ""));
-    return o;
-  }
-};
-WRITE_CLASS_ENCODER(BloomHitSet)
 
 #endif
