+++ src/osd/osd_types.cc
@@ -18,6 +18,7 @@
 
 #include "osd_types.h"
 #include "osd_perf_counters.h"
+#include "BloomHitSet.h"
 
 #include <algorithm>
 #include <list>
